CRPT_6090_N: develop-1.5.5: UI Task to create page for moat and ias

+++ app/scripts/login/login_model.js
@@ -101,7 +101,7 @@ define(['angularAMD'], function (angularAMD) {
             },
 
             checkCookieExpiry: function () {
-                if (!$cookies.get('cdesk_session')) {
+                if (!$location.path().endsWith('_tagValidator') && !$cookies.get('cdesk_session')) {
                     localStorage.clear();
                     updateRedirectUrl($location.$$path);
